Refactored all validation functions into their own files
