+++ infraohjelmointi_api/serializers/ProjectGroupSerializer.py
@@ -4,6 +4,9 @@ from infraohjelmointi_api.serializers import (
     DynamicFieldsModelSerializer,
     ProjectHashtagSerializer,
 )
+from infraohjelmointi_api.validators.ProjectGroupValidators.ProjectsFieldValidator import (
+    ProjectsFieldValidator,
+)
 from rest_framework import serializers
 from rest_framework.validators import UniqueTogetherValidator
 from overrides import override
@@ -15,28 +18,12 @@ from django.shortcuts import get_object_or_404
 class ProjectGroupSerializer(DynamicFieldsModelSerializer):
 
     projects = serializers.ListField(
-        child=serializers.UUIDField(), write_only=True, required=False, allow_empty=True
+        child=serializers.UUIDField(),
+        write_only=True,
+        required=False,
+        allow_empty=True,
+        validators=[ProjectsFieldValidator()],
     )
-
-    def validate_projects(self, projectIds):
-        """
-        Check that project doesn't already belong to a group
-        """
-
-        if projectIds is not None and len(projectIds) > 0:
-            for projectId in projectIds:
-                project = get_object_or_404(Project, pk=projectId)
-                if project.projectGroup is not None:
-                    raise ValidationError(
-                        detail="Project: {} with id: {} already belongs to the group: {} with id: {}".format(
-                            project.name,
-                            projectId,
-                            project.projectGroup.name,
-                            project.projectGroup_id,
-                        ),
-                        code="project_in_group",
-                    )
-        return projectIds
 
     class Meta(BaseMeta):
         model = ProjectGroup
