+++ infraohjelmointi_api/validators/ProjectGroupValidators/__init__.py
@@ -1,0 +1,1 @@
+from .ProjectsFieldValidator import ProjectsFieldValidator
