+++ infraohjelmointi_api/validators/ProjectValidators/ProjectClassValidator.py
@@ -1,0 +1,55 @@
+from rest_framework.exceptions import ValidationError
+
+
+class ProjectClassValidator:
+
+    requires_context = True
+
+    def _is_projectClass_projectLocation_valid(
+        self,
+        projectLocation,
+        projectClass,
+    ) -> bool:
+        if projectClass is None or projectLocation is None:
+            return True
+        if projectClass.parent is None or projectClass.parent.parent is None:
+            return True
+        if (
+            "suurpiiri" in projectClass.name.lower()
+            and len(projectClass.name.split()) == 2
+            and (
+                projectLocation.path.startswith(projectClass.name.split()[0])
+                or projectLocation.path.startswith(projectClass.name.split()[0][:-2])
+            )
+        ):
+            return True
+        elif "suurpiiri" not in projectClass.name.lower():
+            return True
+        else:
+            return False
+
+    def __call__(self, allFields, serializer) -> None:
+        projectClass = allFields.get("projectClass", None)
+        if projectClass is None:
+            return
+        project = serializer.instance
+        projectLocation = allFields.get("projectLocation", None)
+
+        if (
+            projectLocation is None
+            and project is not None
+            and project.projectLocation is not None
+        ):
+            projectLocation = project.projectLocation
+
+        if not self._is_projectClass_projectLocation_valid(
+            projectLocation=projectLocation, projectClass=projectClass
+        ):
+            raise ValidationError(
+                detail="subClass: {} with path: {} cannot have the location: {} under it.".format(
+                    projectClass.name,
+                    projectClass.path,
+                    projectLocation.name,
+                ),
+                code="projectClass_invalid_projectLocation",
+            )
